時刻の取得を機能コンポーネント化
